move test_nonce_info to nonce_info.rs
